Review changes + docs fixes
