--- docs/source/api/geodesic.rst
@@ -1,8 +1,0 @@
-Geodesic module
-===============
-
-This module contains the class, defining a general Geodesic:
-
-.. automodule:: einsteinpy.geodesic
-    :members:
-    :show-inheritance:
