+++ docs/source/api/geodesic/geodesic.rst
@@ -1,0 +1,9 @@
+Geodesic
+========
+
+This module defines a base ``Geodesic`` class, and then 
+extends it to define Time-like and Null-like Geodesics.
+
+.. automodule:: einsteinpy.geodesic.geodesic
+    :members:
+    :show-inheritance:
