+++ docs/source/api/geodesic/geodesic_index.rst
@@ -1,0 +1,9 @@
+Geodesic
+========
+
+This module contains classes for defining Time-like & Null-like Geodesics.
+
+.. toctree::
+    :maxdepth: 2
+
+    geodesic
