+++ docs/source/api/integrators/fantasy.rst
@@ -1,0 +1,10 @@
+FANTASY
+=======
+
+This module uses Forward Mode Automatic Differentiation
+to calculate metric derivatives to machine precision
+leading to stable simulations.
+
+.. automodule:: einsteinpy.integrators.fantasy
+    :members:
+    :show-inheritance:
