+++ docs/source/api/integrators/integrators_index.rst
@@ -1,9 +1,10 @@
-Integrators module
-==================
+Integrators
+===========
 
-This module contains the methods of solving Ordinary Differential Equations.
+This module contains methods for solving Ordinary Differential Equations in Numerical Relativity.
 
 .. toctree::
     :maxdepth: 2
 
     runge_kutta
+    fantasy
