+++ docs/source/api/plotting/geodesic/core.rst
@@ -1,0 +1,8 @@
+Core
+====
+
+This module allows for automatic plotting-backend switching.
+
+.. automodule:: einsteinpy.plotting.geodesic.core
+    :members:
+    :show-inheritance:
