+++ docs/source/api/plotting/geodesic/geodesic_index.rst
@@ -1,0 +1,12 @@
+Plotting Geodesics
+==================
+
+This module contains classes to produce static 
+and interactive, 2D and 3D geodesic plots.
+
+.. toctree::
+    :maxdepth: 2
+
+    core
+    static
+    interactive
