+++ docs/source/api/plotting/geodesic/interactive.rst
@@ -1,0 +1,9 @@
+Interactive Plotting
+====================
+
+This module contains a backend for producing 2D 
+and 3D interactive geodesic plots, using ``plotly``.
+
+.. automodule:: einsteinpy.plotting.geodesic.interactive
+    :members:
+    :show-inheritance:
