+++ docs/source/api/plotting/geodesic/static.rst
@@ -1,0 +1,9 @@
+Static Plotting
+===============
+
+This module contains a backend for producing 2D
+and 3D static geodesic plots, using ``matplotlib``.
+
+.. automodule:: einsteinpy.plotting.geodesic.static
+    :members:
+    :show-inheritance:
