--- docs/source/api/plotting/geodesics/core.rst
@@ -1,8 +1,0 @@
-Core plotting module
-====================
-
-This module contains the advanced backend switching while plotting.
-
-.. automodule:: einsteinpy.plotting.geodesic.core
-    :members:
-    :show-inheritance:
