--- docs/source/api/plotting/geodesics/geodesics_index.rst
@@ -1,10 +1,0 @@
-Geodesics Plotting module
-=========================
-
-This module contains Classes to  plot geodesics.
-
-.. toctree::
-    :maxdepth: 2
-
-    core
-    interactive
