--- docs/source/api/plotting/geodesics/interactive.rst
@@ -1,8 +1,0 @@
-Interactive plotting module
-===========================
-
-This module contains interactive plotting backend.
-
-.. automodule:: einsteinpy.plotting.geodesic.interactive
-    :members:
-    :show-inheritance:
