+++ docs/source/api/utils/dual.rst
@@ -1,0 +1,9 @@
+Dual Number
+===========
+
+This module adds support for Dual Numbers and 
+Forward Mode Auto Differentiation.
+
+.. automodule:: einsteinpy.utils.dual
+    :members:
+    :show-inheritance:
