+++ docs/source/api/utils/utils_index.rst
@@ -1,9 +1,11 @@
-Utils module
-============
+Utils
+=====
 
-The common utilities of the project are included in this module.
+This module stores the common utilities of the project, such as a ``DualNumber``
+class, that defines dual numbers, used for Forward Mode Auto Differentiation.
 
 .. toctree::
     :maxdepth: 2
 
+    dual
     scalar_factor
