+++ setup.cfg
@@ -43,7 +43,6 @@ install_requires =
         plotly>=4.0
         sympy>=1.1
         numba>=0.46,!=0.49.0 ; implementation_name=='cpython'
-        einsteinpy-geodesics
 include_package_data = True
 python_requires = >=3.7
 
@@ -68,7 +67,6 @@ dev =
         ipywidgets
         tox
         jupyter_sphinx
-        einsteinpy-geodesics
 
 [tool:pytest]
 norecursedirs =
