+++ src/einsteinpy/integrators/utils.py
@@ -4,7 +4,7 @@ Utilities for Integration Module
 """
 import numpy as np
 
-from einsteinpy.misc.dual import Dual, _deriv, _diff_g, _jacobian_g
+from einsteinpy.utils.dual import DualNumber, _deriv, _diff_g, _jacobian_g
 
 
 def _PartHamFlow(g, g_prms, q, p, wrt):
@@ -37,7 +37,7 @@ def _PartHamFlow(g, g_prms, q, p, wrt):
     .. [1] Christian, Pierre and Chan, Chi-Kwan;
         "FANTASY: User-Friendly Symplectic Geodesic Integrator
         for Arbitrary Metrics with Automatic Differentiation";
-        `arXiv:2010.02237 <https://arxiv.org/abs/2010.02237>`_
+        `arXiv:2010.02237 <https://arxiv.org/abs/2010.02237>`__
 
     """
     return _jacobian_g(g, g_prms, q, wrt) @ p @ p
@@ -76,7 +76,7 @@ def _flow_A(g, g_prms, q1, p1, q2, p2, delta=0.5):
     .. [1] Christian, Pierre and Chan, Chi-Kwan;
         "FANTASY: User-Friendly Symplectic Geodesic Integrator
         for Arbitrary Metrics with Automatic Differentiation";
-        `arXiv:2010.02237 <https://arxiv.org/abs/2010.02237>`_
+        `arXiv:2010.02237 <https://arxiv.org/abs/2010.02237>`__
 
     """
     dH1 = [0.5 * (_PartHamFlow(g, g_prms, q1, p2, i)) for i in range(4)]
@@ -123,7 +123,7 @@ def _flow_B(g, g_prms, q1, p1, q2, p2, delta=0.5):
     .. [1] Christian, Pierre and Chan, Chi-Kwan;
         "FANTASY: User-Friendly Symplectic Geodesic Integrator
         for Arbitrary Metrics with Automatic Differentiation";
-        `arXiv:2010.02237 <https://arxiv.org/abs/2010.02237>`_
+        `arXiv:2010.02237 <https://arxiv.org/abs/2010.02237>`__
 
     """
     dH2 = [
@@ -151,7 +151,7 @@ def _flow_B(g, g_prms, q1, p1, q2, p2, delta=0.5):
 
 def _flow_mixed(q1, p1, q2, p2, delta=0.5, omega=1.0):
     """
-    Mixed flow of Hamiltonian, :math:`\tilde{H}`
+    Mixed flow of Hamiltonian, :math:`\\tilde{H}`
 
     Parameters
     ----------
@@ -173,14 +173,14 @@ def _flow_mixed(q1, p1, q2, p2, delta=0.5, omega=1.0):
     Returns
     -------
     float
-        Hamiltonian Flow for :math:`\tilde{H}`
+        Hamiltonian Flow for :math:`\\tilde{H}`
 
     References
     ----------
     .. [1] Christian, Pierre and Chan, Chi-Kwan;
         "FANTASY: User-Friendly Symplectic Geodesic Integrator
         for Arbitrary Metrics with Automatic Differentiation";
-        `arXiv:2010.02237 <https://arxiv.org/abs/2010.02237>`_
+        `arXiv:2010.02237 <https://arxiv.org/abs/2010.02237>`__
 
     """
     q_sum = q1 + q2
