--- src/einsteinpy/misc/__init__.py
@@ -1,1 +1,0 @@
-from .dual import Dual, _deriv, _diff_g, _jacobian_g
