+++ src/einsteinpy/utils/__init__.py
@@ -1,1 +1,2 @@
+from .dual import DualNumber, _deriv, _diff_g, _jacobian_g
 from .scalar_factor import scalar_factor, scalar_factor_derivative
